Let's make this an engine
Engines let us interact with the Rails application and do interesting
things like providing javascript for the Rails 3.1 asset pipeline.

After a long time of letting this collect dust, I finally have time to
work on the project it was intended to be part of. With Rails 3.1 and
the asset pipeline, it makes sense to convert this to an engine and
provide some "boilerplate" Facebook javascript initialization code to
the asset pipeline. This is intended to support Rails 3.1, while still
maintaining compatibility with Rails 3.0. It will require some testing
before I'm ready to merge into master and release a new gem.